Move RefMetadata implementation for 'virtual-branches.toml' to `but-graph`.
`but-core` really wants to have it, but it has dependencies that the core may never use.
Thus, it must be in `graph` for now until it can be removed entirely.
