+++ Cargo.toml
@@ -80,6 +80,8 @@ but-graph = { path = "crates/but-graph" }
 but-action = { path = "crates/but-action" }
 but-status = { path = "crates/but-status" }
 git2-hooks = { version = "0.5.0" }
+itertools = "0.14.0"
+
 
 [profile.release]
 codegen-units = 1 # Compile crates one after another so the compiler can optimize better
