+++ crates/but-action/Cargo.toml
@@ -31,7 +31,7 @@ gitbutler-oxidize.workspace = true
 gitbutler-project.workspace = true
 gitbutler-reference.workspace = true
 gitbutler-serde.workspace = true
-but-core.workspace = true
 but-db.workspace = true
+but-graph.workspace = true
 but-workspace.workspace = true
 but-hunk-assignment.workspace = true
