+++ crates/but-graph/tests/graph/main.rs
@@ -320,3 +320,5 @@ mod utils {
     }
 }
 pub use utils::graph_tree;
+
+mod ref_metadata_legacy;
