+++ crates/but-hunk-dependency/Cargo.toml
@@ -10,7 +10,7 @@ doctest = false
 
 [dependencies]
 anyhow = "1.0.98"
-itertools = "0.14.0"
+itertools.workspace = true
 serde.workspace = true
 gix = { workspace = true, features = [] }
 but-core.workspace = true
@@ -30,6 +30,5 @@ gitbutler-stack.workspace = true
 gix-testtools.workspace = true
 gitbutler-testsupport.workspace = true
 gitbutler-oxidize.workspace = true
-itertools = "0.14.0"
 insta.workspace = true
 serde_json = "1.0.138"
