+++ crates/but-workspace/tests/workspace/main.rs
@@ -1,6 +1,5 @@
 mod branch_details;
 mod commit_engine;
 mod ref_info;
-mod ref_metadata;
 mod tree_manipulation;
 mod utils;
