+++ crates/but/Cargo.toml
@@ -26,11 +26,8 @@ gitbutler-project.workspace = true
 gix.workspace = true
 but-core.workspace = true
 but-action.workspace = true
+but-graph.workspace = true
 but-workspace.workspace = true
 but-settings.workspace = true
-but-hunk-dependency.workspace = true
 gitbutler-command-context.workspace = true
 serde_json = "1.0.140"
-dirs-next = "2.0.0"
-gitbutler-branch-actions.workspace = true
-gitbutler-branch.workspace = true
